backend/requirements.txt
pandas
scikit-learn
joblib
flask
pymysql
PyJWT
bcrypt
cryptography